requirements.txt
asgiref==3.11.0
Django==6.0.1
sqlparse==0.5.5
tzdata==2025.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
black==26.1.0
flake8==7.3.0
isort==8.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
